+++ .travis.yml
@@ -7,6 +7,5 @@ before_install:
 - docker ps -a
 
 script:  
-- docker exec -it travisImage pytest /srv/www/todo_app/tests/function_tests.py
-
+- docker exec -it travisImage pytest /srv/www/todo_app/tests/
