--- todo_app/tests/tdd_tests.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-pytest function_tests.py
-
